! fixed all remaining warnings but 3 (those 3 are 64bit only... and they shouldn't matter.. but don't know how to "fix" them) * updated project files accordingly
+! redesigned StopWatch dialog, it now really stops (time doesn't run), laps also only measure when running etc. It's now also possible to reset while running. Performance also improved (StopWatch.c, main.c, tClock.rc, resource.h)

+++ Source/Clock/PageMisc.c
@@ -146,7 +146,7 @@ void OnApply(HWND hDlg)   //----------------------------------------------------
 	SetMyRegLong("Calendar","ShowWeekNums", IsDlgButtonChecked(hDlg,IDCB_SHOWWEEKNUMS));
 	SetMyRegLong("Calendar","ShowDayOfYear", IsDlgButtonChecked(hDlg,IDCB_SHOW_DOY));
 	SetMyRegLong("Calendar","CalendarTopMost", IsDlgButtonChecked(hDlg,IDCB_CALTOPMOST));
-	SetMyRegLong("Calendar", "ViewMonths", SendDlgItemMessage(hDlg,IDC_CALMONTHSPIN,UDM_GETPOS,0,0));
+	SetMyRegLong("Calendar", "ViewMonths", (DWORD)SendDlgItemMessage(hDlg,IDC_CALMONTHSPIN,UDM_GETPOS,0,0));
 	SetMyRegLong("Desktop","Transparent2kIconText", IsDlgButtonChecked(hDlg,IDCB_TRANS2KICONS));
 	
 	GetDlgItemText(hDlg, IDC_FIRSTWEEK, szWeek, 8);
